Fix system tests
The original system test was assuming that the daemon would start
scanning by itself.  Now, we need set-up a job on the server and pass
the proper credentials to the daemon for it to get the scanning
instruction.  The rest is the same (wait for the scanning to finish and
check the times at which the images appeared...)

I decided to experiment a bit with testing in production environment
("staging" actually but the idea is the same): instead of running an
isolated copy of scannomatic to run the tests, I use the existing
staging environement.  There are a few advantages:
- simpler and faster test setup
- tests are done in real condidions (e.g. with password authentication)
- tests are run on a Pi, which can't run scannomatic easily

There are also a few potential problems:
- there's only one scanner, which means there can't be concurrent test
  runs. This will be fixed once we can "register" scanners in some way
- there's a small chance that the test fails because scanomatic is
  redeployed.  If this is a problem, we can tell jenkins to coordinate
  job correctly in order to avoid it.